Move OpenShift platform classes into own directory
OpenShift platform detection classes and auth manager are moved into
their own openshift directory, with the auth manager made into its own
public non-nested class for ease of specification with
CONTAINER_JFR_AUTH_MANAGER environment variable
