+++ src/main/java/com/redhat/rhjmc/containerjfr/platform/internal/DefaultPlatformClient.java
@@ -40,6 +40,6 @@ class DefaultPlatformClient implements PlatformClient {
 
     @Override
     public AuthManager getAuthManager() {
-        return new NoopAuthManager(log);
+        return new NoopAuthManager();
     }
 }
