+++ src/main/java/com/redhat/rhjmc/containerjfr/platform/internal/KubeEnvPlatformClient.java
@@ -46,6 +46,6 @@ class KubeEnvPlatformClient implements PlatformClient {
 
     @Override
     public AuthManager getAuthManager() {
-        return new NoopAuthManager(logger);
+        return new NoopAuthManager();
     }
 }
